# apps/portal/registry/openapi/portal-api.openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilityAvailability": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "authority": {
            "$ref": "#/components/schemas/AvailabilityAuthority"
          },
          "checked_at": {
            "format": "date-time",
            "title": "Checked At",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "provenance": {
            "$ref": "#/components/schemas/AvailabilityProvenance"
          },
          "reason_code": {
            "anyOf": [
              {
                "enum": [
                  "CAPABILITY_NOT_IMPLEMENTED",
                  "UPSTREAM_UNAVAILABLE",
                  "UPSTREAM_TIMEOUT",
                  "INCOMPATIBLE_CONTRACT",
                  "SOURCE_DATA_UNAVAILABLE",
                  "LOCAL_ONLY_STATE",
                  "PERMISSION_DENIED",
                  "STALE_OBSERVATION",
                  "PARTIAL_SOURCE_FAILURE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "stale_after_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale After Seconds"
          },
          "state": {
            "enum": [
              "available",
              "unavailable",
              "degraded",
              "stale",
              "denied",
              "commissioned"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "reason_code",
          "detail",
          "retryable",
          "checked_at",
          "as_of",
          "stale_after_seconds",
          "authority",
          "provenance"
        ],
        "title": "CapabilityAvailability",
        "type": "object"
      },
...
      "ConcernDefinition": {
        "additionalProperties": false,
        "properties": {
          "activation_gate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activation Gate"
          },
          "category": {
            "enum": [
              "PRODUCT_DECISION",
              "AUTHORIZATION",
              "SOURCE_OF_TRUTH",
              "BACKEND_CONTRACT",
              "STATE_MACHINE",
              "RESILIENCE_STATE",
              "QUANT_SEMANTICS",
              "PERFORMANCE",
              "AUDIT_LINEAGE",
              "SECURITY",
              "ACCESSIBILITY",
              "TEST_EVIDENCE",
              "DEPENDENCY",
              "ACTIVATION_GATE"
            ],
            "title": "Category",
            "type": "string"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "feature_ids": {
            "items": {
              "type": "string"
            },
            "title": "Feature Ids",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reviewed_at": {
            "title": "Reviewed At",
            "type": "string"
          },
          "screen_ids": {
            "items": {
              "type": "string"
            },
            "title": "Screen Ids",
            "type": "array"
          },
          "severity": {
            "enum": [
              "BLOCKING",
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Severity",
            "type": "string"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "PARTIAL",
              "VERIFIED_CURRENT",
              "BLOCKED",
              "NOT_APPLICABLE"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_ids": {
            "items": {
              "type": "string"
            },
            "title": "Task Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "category",
          "status",
          "severity",
          "statement",
          "feature_ids",
          "screen_ids",
          "evidence_refs",
          "task_ids",
          "activation_gate",
          "reviewed_at"
        ],
        "title": "ConcernDefinition",
        "type": "object"
      },
...
      "FeatureGroupDefinition": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "enum": [
              "command",
              "research",
              "backtests",
              "deployments",
              "data_ops",
              "planning",
              "administration"
            ],
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "label",
          "order"
        ],
        "title": "FeatureGroupDefinition",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LifecycleStageDefinition": {
        "additionalProperties": false,
        "description": "Roll-up of ``primary_persona`` across the stage's feature screens.\n\n``personas`` is derived at projection time (never stored in the source\nregistry document); stages whose features have no screens yet carry an\nempty tuple so the UI can filter/hide them on persona mismatch.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "feature_ids": {
            "items": {
              "type": "string"
            },
            "title": "Feature Ids",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "maturity": {
            "enum": [
              "AVAILABLE",
              "PROTOTYPE",
              "COMMISSIONED",
              "BLOCKED",
              "HIDDEN",
              "DEPRECATED"
            ],
            "title": "Maturity",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          },
          "personas": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Personas",
            "type": "array"
          }
        },
        "required": [
          "id",
          "label",
          "order",
          "feature_ids",
          "maturity",
          "description"
        ],
        "title": "LifecycleStageDefinition",
        "type": "object"
      },
...
      "PortalErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/PortalErrorDetail"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "title": "PortalErrorResponse",
        "type": "object"
      },
      "PortalFeatureDefinition": {
        "additionalProperties": false,
        "properties": {
          "activation_gate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activation Gate"
          },
          "canonical_route": {
            "title": "Canonical Route",
            "type": "string"
          },
          "concern_ids": {
            "items": {
              "type": "string"
            },
            "title": "Concern Ids",
            "type": "array"
          },
          "data_mode": {
            "enum": [
              "REAL",
              "FIXTURE",
              "STATIC_PREVIEW",
              "NONE"
            ],
            "title": "Data Mode",
            "type": "string"
          },
          "default_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Task Id"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "environments": {
            "items": {
              "enum": [
                "local",
                "research",
                "paper",
                "sandbox",
                "live"
              ],
              "type": "string"
            },
            "title": "Environments",
            "type": "array"
          },
          "group": {
            "enum": [
              "command",
              "research",
              "backtests",
              "deployments",
              "data_ops",
              "planning",
              "administration"
            ],
            "title": "Group",
            "type": "string"
          },
          "hidden_for_roles": {
            "items": {
              "type": "string"
            },
            "title": "Hidden For Roles",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "legacy_routes": {
            "items": {
              "type": "string"
            },
            "title": "Legacy Routes",
            "type": "array"
          },
          "lifecycle_stage_ids": {
            "items": {
              "type": "string"
            },
            "title": "Lifecycle Stage Ids",
            "type": "array"
          },
          "maturity": {
            "enum": [
              "AVAILABLE",
              "PROTOTYPE",
              "COMMISSIONED",
              "BLOCKED",
              "HIDDEN",
              "DEPRECATED"
            ],
            "title": "Maturity",
            "type": "string"
          },
          "navigation": {
            "$ref": "#/components/schemas/NavigationDefinition"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "prototype_frame_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prototype Frame Id"
          },
          "roadmap_epic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roadmap Epic Id"
          },
          "screen_ids": {
            "items": {
              "type": "string"
            },
            "title": "Screen Ids",
            "type": "array"
          },
          "source_module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Module"
          },
          "summary_source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Summary Source Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "group",
          "label",
          "description",
          "canonical_route",
          "legacy_routes",
          "maturity",
          "data_mode",
          "permissions",
          "environments",
          "source_module",
          "prototype_frame_id",
          "roadmap_epic_id",
          "default_task_id",
          "screen_ids",
          "concern_ids",
          "lifecycle_stage_ids",
          "summary_source_ids",
          "hidden_for_roles",
          "activation_gate",
          "navigation"
        ],
        "title": "PortalFeatureDefinition",
        "type": "object"
      },
      "PortalLinkEntry": {
        "additionalProperties": false,
        "properties": {
          "activation_gate": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activation Gate"
          },
          "concern_id": {
            "anyOf": [
              {
                "pattern": "^[A-Z][A-Z0-9_]{2,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concern Id"
          },
          "feature_id": {
            "anyOf": [
              {
                "pattern": "^[A-Z][A-Z0-9_]{2,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Id"
          },
          "figma_frame_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Figma Frame Id"
          },
          "id": {
            "pattern": "^[A-Z][A-Z0-9_]{2,63}$",
            "title": "Id",
            "type": "string"
          },
          "planning_task_ids": {
            "items": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,119}$",
              "type": "string"
            },
            "maxItems": 50,
            "title": "Planning Task Ids",
            "type": "array"
          },
          "prototype_route": {
            "anyOf": [
              {
                "maxLength": 240,
                "pattern": "^/",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prototype Route"
          },
          "repository_scope": {
            "items": {
              "pattern": "^[a-z][a-z0-9-]{1,63}$",
              "type": "string"
            },
            "maxItems": 16,
            "title": "Repository Scope",
            "type": "array"
          },
          "roadmap_epic_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roadmap Epic Id"
          },
          "screen_id": {
            "anyOf": [
              {
                "pattern": "^[A-Z][A-Z0-9_]{2,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screen Id"
          }
        },
        "required": [
          "id",
          "feature_id",
          "screen_id",
          "concern_id",
          "roadmap_epic_id",
          "planning_task_ids",
          "figma_frame_id",
          "repository_scope"
        ],
        "title": "PortalLinkEntry",
        "type": "object"
      },
      "PortalLinksDocument": {
        "additionalProperties": false,
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/PortalLinkEntry"
            },
            "maxItems": 128,
            "title": "Entries",
            "type": "array"
          },
          "integrity": {
            "$ref": "#/components/schemas/PortalLinksIntegrity"
          },
          "links_revision": {
            "minimum": 1.0,
            "title": "Links Revision",
            "type": "integer"
          },
          "reviewed_at": {
            "title": "Reviewed At",
            "type": "string"
          },
          "schema_version": {
            "const": "portal.links.v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "links_revision",
          "reviewed_at",
          "entries",
          "integrity"
        ],
        "title": "PortalLinksDocument",
        "type": "object"
      },
      "PortalLinksIntegrity": {
        "additionalProperties": false,
        "properties": {
          "concerns_linked": {
            "minimum": 0.0,
            "title": "Concerns Linked",
            "type": "integer"
          },
          "dangling_links": {
            "default": 0,
            "title": "Dangling Links",
            "type": "integer"
          },
          "features_linked": {
            "minimum": 0.0,
            "title": "Features Linked",
            "type": "integer"
          },
          "planning_tasks_referenced": {
            "minimum": 0.0,
            "title": "Planning Tasks Referenced",
            "type": "integer"
          },
          "screens_linked": {
            "minimum": 0.0,
            "title": "Screens Linked",
            "type": "integer"
          },
          "status": {
            "const": "valid",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "features_linked",
          "screens_linked",
          "concerns_linked",
          "planning_tasks_referenced"
        ],
        "title": "PortalLinksIntegrity",
        "type": "object"
      },
      "PortalRegistryDocument": {
        "additionalProperties": false,
        "properties": {
          "concerns": {
            "items": {
              "$ref": "#/components/schemas/ConcernDefinition"
            },
            "title": "Concerns",
            "type": "array"
          },
          "content_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Content Digest",
            "type": "string"
          },
          "feature_groups": {
            "items": {
              "$ref": "#/components/schemas/FeatureGroupDefinition"
            },
            "title": "Feature Groups",
            "type": "array"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/PortalFeatureDefinition"
            },
            "title": "Features",
            "type": "array"
          },
          "lifecycle_stages": {
            "items": {
              "$ref": "#/components/schemas/LifecycleStageDefinition"
            },
            "title": "Lifecycle Stages",
            "type": "array"
          },
          "registry_id": {
            "const": "portal-default",
            "title": "Registry Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "schema_version": {
            "const": "portal.registry.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "screens": {
            "items": {
              "$ref": "#/components/schemas/ScreenContract"
            },
            "title": "Screens",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "registry_id",
          "revision",
          "feature_groups",
          "lifecycle_stages",
          "features",
          "screens",
          "concerns",
          "content_digest"
        ],
        "title": "PortalRegistryDocument",
        "type": "object"
      },
...
      "PortalSummaryV1": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "environment": {
            "enum": [
              "local",
              "research",
              "paper",
              "sandbox",
              "live"
            ],
            "title": "Environment",
            "type": "string"
          },
          "overall_availability": {
            "$ref": "#/components/schemas/CapabilityAvailability"
          },
          "priority_items": {
            "items": {
              "$ref": "#/components/schemas/PriorityItem"
            },
            "maxItems": 50,
            "title": "Priority Items",
            "type": "array"
          },
          "registry_counts": {
            "$ref": "#/components/schemas/RegistryCounts"
          },
          "registry_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Registry Digest",
            "type": "string"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "schema_version": {
            "const": "portal.summary.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/PortalSummarySection"
            },
            "maxItems": 32,
            "title": "Sections",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "registry_digest",
          "environment",
          "requested_at",
          "completed_at",
          "overall_availability",
          "registry_counts",
          "sections",
          "priority_items"
        ],
        "title": "PortalSummaryV1",
        "type": "object"
      },
...
      "PriorityItem": {
        "additionalProperties": false,
        "properties": {
          "authority": {
            "title": "Authority",
            "type": "string"
          },
          "evidence_digest": {
            "anyOf": [
              {
                "pattern": "^sha256:[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Digest"
          },
          "feature_id": {
            "title": "Feature Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "route": {
            "maxLength": 240,
            "pattern": "^/",
            "title": "Route",
            "type": "string"
          },
          "severity": {
            "enum": [
              "critical",
              "warning",
              "info"
            ],
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "RUN_FAILED",
              "HISTORICAL_DATA_UNAVAILABLE",
              "REGISTRY_BLOCKING_CONCERN"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "severity",
          "title",
          "feature_id",
          "resource_id",
          "observed_at",
          "authority",
          "route",
          "evidence_digest"
        ],
        "title": "PriorityItem",
        "type": "object"
      },
...
      "RegistryCounts": {
        "additionalProperties": false,
        "properties": {
          "blocking_concerns": {
            "minimum": 0.0,
            "title": "Blocking Concerns",
            "type": "integer"
          },
          "by_maturity": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Maturity",
            "type": "object"
          }
        },
        "required": [
          "by_maturity",
          "blocking_concerns"
        ],
        "title": "RegistryCounts",
        "type": "object"
      },
...
      "ScreenContract": {
        "additionalProperties": false,
        "properties": {
          "activation_gate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activation Gate"
          },
          "backend_dependency_ids": {
            "items": {
              "type": "string"
            },
            "title": "Backend Dependency Ids",
            "type": "array"
          },
          "concern_ids": {
            "items": {
              "type": "string"
            },
            "title": "Concern Ids",
            "type": "array"
          },
          "contract_revision": {
            "title": "Contract Revision",
            "type": "integer"
          },
          "data_mode": {
            "enum": [
              "REAL",
              "FIXTURE",
              "STATIC_PREVIEW",
              "NONE"
            ],
            "title": "Data Mode",
            "type": "string"
          },
          "feature_id": {
            "title": "Feature Id",
            "type": "string"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/ScreenInputDefinition"
            },
            "title": "Inputs",
            "type": "array"
          },
          "maturity": {
            "enum": [
              "AVAILABLE",
              "PROTOTYPE",
              "COMMISSIONED",
              "BLOCKED",
              "HIDDEN",
              "DEPRECATED"
            ],
            "title": "Maturity",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "primary_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Action"
          },
          "primary_decision": {
            "title": "Primary Decision",
            "type": "string"
          },
          "primary_persona": {
            "title": "Primary Persona",
            "type": "string"
          },
          "related_repositories": {
            "items": {
              "type": "string"
            },
            "title": "Related Repositories",
            "type": "array"
          },
          "related_task_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Task Ids",
            "type": "array"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "screen_id": {
            "title": "Screen Id",
            "type": "string"
          }
        },
        "required": [
          "screen_id",
          "contract_revision",
          "feature_id",
          "maturity",
          "data_mode",
          "route",
          "primary_persona",
          "primary_decision",
          "primary_action",
          "permissions",
          "inputs",
          "backend_dependency_ids",
          "concern_ids",
          "related_repositories",
          "related_task_ids",
          "activation_gate"
        ],
        "title": "ScreenContract",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QuantBT Backtest Portal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/portal/links": {
      "get": {
        "operationId": "portal_links_api_v1_portal_links_get",
        "parameters": [
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalLinksDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Links content has not changed."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Portal Links",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/portal/registry": {
      "get": {
        "operationId": "portal_registry_api_v1_portal_registry_get",
        "parameters": [
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalRegistryDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Registry content has not changed."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Portal Registry",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/portal/summary": {
      "get": {
        "operationId": "portal_summary_api_v1_portal_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalSummaryV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalErrorResponse"
                }
              }
            },
            "description": "Summary internal contract failure."
          }
        },
        "summary": "Portal Summary",
        "tags": [
          "portal"
        ]
      }
    }
  }
}